Job updates for upgrades
While both jobs technically now both create. I have decided to leave CrawlMediaPageJob as is so I can focus on creating a new feature of adding in a service.

Didn't seem to make sense to use the ProfilePage as a way of creating meta data, it should just save from the API directly.