Feature: Catalogue Service
Major refactor of all the complicated tasks/jobs and entirety of catalogue metadata fetching.

Everything is now in a nice tidy Service from which we can call from anywhere.

Each endpoint/request from the service is executed via hydrate functions that call their assigned objects and relations to the main Catalogue object.

New graphql field to show Folder titles as I think that is easier to read in the CMS than `xyz-imdbID` and keeps it separate from any URL routing Silverstripe provides.

By default imported or CMS created catalogue objects will now be marked as "incomplete" this allows us for a separate view in the CMS to see which fetching or incomplete records need attention. This is not applicable for records created via the front end forms as that always hits OMDB Api before its inserted.

Removed metadata and poster folders all that served to do was complicate any hydrate methods and  bloating the folder structure. We just now use
`catalogue-metadata/title-of-movie-imdbid`.

Finally I've added some tests to how the metadata folder structure is created but obviously this needs its own feature to dive into coverage for everything related.